added readme.md detailing project structure added config to run the EDDN message handlers async
