+++ src/main/java/io/eddb/eddb2backend/application/controller/station/StationController.java
@@ -1,14 +1,16 @@
 package io.eddb.eddb2backend.application.controller.station;
 
 import io.eddb.eddb2backend.application.dto.station.GetStationResponse;
+import io.eddb.eddb2backend.application.mapper.StationMapper;
 import io.eddb.eddb2backend.application.usecase.GetStationUsecase;
-import java.util.Collection;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.Collection;
 
 @RestController
 @RequestMapping("/api/stations")
@@ -20,7 +22,7 @@ public class StationController {
     @GetMapping("/{id}")
     public ResponseEntity<GetStationResponse> getStation(@PathVariable Long id) {
         return getStationUsecase.getById(id)
-                .map(GetStationResponse.Mapper::map)
+                .map(StationMapper::map)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
     }
@@ -29,7 +31,7 @@ public class StationController {
     public ResponseEntity<Collection<GetStationResponse>> getAllStations() {
         var responseList = getStationUsecase.getAll()
                 .stream()
-                .map(GetStationResponse.Mapper::map)
+                .map(StationMapper::map)
                 .toList();
         return ResponseEntity.ok(responseList);
     }
