+++ src/main/java/io/eddb/eddb2backend/configuration/ZeromqConfig.java
@@ -1,16 +1,13 @@
 package io.eddb.eddb2backend.configuration;
 
-import io.eddb.eddb2backend.infrastructure.external.EddnMessageHandler;
-import java.io.ByteArrayOutputStream;
-import java.nio.charset.StandardCharsets;
-import java.util.zip.DataFormatException;
-import java.util.zip.Inflater;
+import io.eddb.eddb2backend.infrastructure.eddn.EddnMessageHandler;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.task.TaskExecutor;
 import org.springframework.integration.annotation.ServiceActivator;
 import org.springframework.integration.config.EnableIntegration;
 import org.springframework.integration.zeromq.channel.ZeroMqChannel;
-import org.springframework.messaging.MessageHandler;
 import org.zeromq.ZContext;
 
 @Configuration
@@ -23,17 +20,17 @@ public class ZeromqConfig {
     }
 
     @Bean(name = "zeroMqChannel")
-    ZeroMqChannel zeroMqPubSubChannel(ZContext context) {
+    public ZeroMqChannel zeroMqPubSubChannel(ZContext context, @Value("${eddn.uri}") String eddnUri) {
         ZeroMqChannel channel = new ZeroMqChannel(context, true);
-        channel.setConnectUrl("tcp://eddn.edcd.io:9500:9500");
+        channel.setConnectUrl(eddnUri);
         return channel;
     }
 
 
     @Bean
     @ServiceActivator(inputChannel = "zeroMqChannel")
-    public EddnMessageHandler eddnMessageHandler() {
-        return new EddnMessageHandler();
+    public EddnMessageHandler eddnMessageHandler(TaskExecutor taskExecutor) {
+        return new EddnMessageHandler(taskExecutor);
     }
 
 }
