+++ src/main/resources/application.properties
@@ -1,1 +1,5 @@
-
+eddn.uri=tcp://eddn.edcd.io:9500:9500
+eddn.zeromq.executor.corePoolSize=10
+eddn.zeromq.executor.maxPoolSize=20
+eddn.zeromq.executor.queueCapacity=100
+eddn.zeromq.executor.threadNamePrefix=EddnZeroMQListener-500
